// frontend/src/client/models/ReportOutDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ReportOutDto = {
    title: string;
    description: string;
    authorName: string;
    authorAge: number;
    id: string;
    createdAt: string;
    updatedAt: string;
    files: Array<string>;
};